// Front End/src/app/models/terminal.ts
export class Terminal {
    terminalId: Number;
    terminalCode: String;
    locationId: Number;
    trmlSrvType: String;
    windowNumber: Number;
    validStatus: Number;
    adminId: String;
    terminalIp: String;
    trmlSrvNumber: Number;
    commissionDate: Date;
    updatedBy: String;
    updatedOn: Date;
}